Convert net.js to TypeScript
